Move controller to internal/pkg and main.go to cmd
This follows a more default golang project layout by moving the binary
into `cmd/seccomp-operator` and the library into `internal/pkg`.
