+++ cmd/seccomp-operator/main.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/klog/klogr"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/saschagrunert/seccomp-operator/controllers/profile"
+	"github.com/saschagrunert/seccomp-operator/internal/pkg/controllers/profile"
 )
 
 const (
